+++ src/Equinox.Core/Category.fs
@@ -16,10 +16,10 @@ type ICategory<'event, 'state, 'context> =
     /// - Conflict: signifies the sync failed, and the proposed decision hence needs to be reconsidered in light of the supplied conflicting Stream State
     /// NB the central precondition upon which the sync is predicated is that the stream has not diverged from the `originState` represented by `token`
     ///    where the precondition is not met, the SyncResult.Conflict bears a [lazy] async result (in a specific manner optimal for the store)
-    abstract TrySync: log: ILogger * categoryName: string * streamId: string * streamName: string * 'context
-                      * maybeInit: (CancellationToken -> Task<unit>) voption
-                      * originToken: StreamToken * originState: 'state * events: 'event[]
-                      * CancellationToken -> Task<SyncResult<'state>>
+    abstract Sync: log: ILogger * categoryName: string * streamId: string * streamName: string * 'context
+                   * maybeInit: (CancellationToken -> Task<unit>) voption
+                   * originToken: StreamToken * originState: 'state * events: 'event[]
+                   * CancellationToken -> Task<SyncResult<'state>>
 
 // Low level stream impl, used by Store-specific Category types that layer policies such as Caching in
 namespace Equinox
@@ -35,21 +35,22 @@ type Category<'event, 'state, 'context>
     (   resolveInner: string -> string-> struct (Core.ICategory<'event, 'state, 'context> * string * (CancellationToken -> Task<unit>) voption),
         empty: struct (Core.StreamToken * 'state)) =
     /// Provides access to the low level store operations used for Loading and/or Syncing updates via the Decider
-    /// (Normal usage is via the adjacent `module Decider` / `DeciderExtensions.Resolve` helpers)
+    /// (Normal usage is via the adjacent `module Decider` / `Factory.Resolve` helpers)
     member _.Stream(log: Serilog.ILogger, context: 'context, categoryName, streamId) =
         let struct (inner, streamName, init) = resolveInner categoryName streamId
         { new Core.IStream<'event, 'state> with
+            member _.Name = streamName
             member _.LoadEmpty() =
                 empty
             member _.Load(maxAge, requireLeader, ct) = task {
                 use act = source.StartActivity("Load", ActivityKind.Client)
                 if act <> null then act.AddStream(categoryName, streamId, streamName).AddLeader(requireLeader).AddStale(maxAge) |> ignore
                 return! inner.Load(log, categoryName, streamId, streamName, maxAge, requireLeader, ct) }
-            member _.TrySync(attempt, (token, originState), events, ct) = task {
-                use act = source.StartActivity("TrySync", ActivityKind.Client)
+            member _.Sync(attempt, token, originState, events, ct) = task {
+                use act = source.StartActivity("Sync", ActivityKind.Client)
                 if act <> null then act.AddStream(categoryName, streamId, streamName).AddSyncAttempt(attempt) |> ignore
                 let log = if attempt = 1 then log else log.ForContext("attempts", attempt)
-                return! inner.TrySync(log, categoryName, streamId, streamName, context, init, token, originState, events, ct) } }
+                return! inner.Sync(log, categoryName, streamId, streamName, context, init, token, originState, events, ct) } }
 
 [<AbstractClass; Sealed>]
 type private Stream private () =