+++ src/Equinox.MemoryStore/MemoryStore.fs
@@ -4,7 +4,6 @@
 namespace Equinox.MemoryStore
 
 open Equinox.Core
-open System.Threading.Tasks
 
 /// Each Sync that results in an append to the store is notified via the Store's `Committed` event
 type Commit<'Format> = (struct (FsCodec.StreamName * FsCodec.ITimelineEvent<'Format>[]))
@@ -16,11 +15,11 @@ type VolatileStore<'Format> private (lock) =
     let streams = System.Collections.Concurrent.ConcurrentDictionary<string, FsCodec.ITimelineEvent<'Format>[]>()
 
     let seedStream _streamName struct (_expectedCount, events) = events
-    let updateValue _streamName (currentValue : FsCodec.ITimelineEvent<'Format>[]) struct (expectedCount, events) =
+    let updateValue _streamName (currentValue: FsCodec.ITimelineEvent<'Format>[]) struct (expectedCount, events) =
         if currentValue.Length <> expectedCount then currentValue
         // note we don't publish here, as this function can be potentially invoked multiple times where there is a race
         else Array.append currentValue events
-    let trySync streamName expectedCount events : struct (bool * FsCodec.ITimelineEvent<'Format>[]) =
+    let trySync streamName expectedCount events: struct (bool * FsCodec.ITimelineEvent<'Format>[]) =
         let res = streams.AddOrUpdate(streamName, seedStream, updateValue, (expectedCount, events))
         (obj.ReferenceEquals(Array.last res, Array.last events), res)
 
@@ -29,7 +28,7 @@ type VolatileStore<'Format> private (lock) =
     /// Commits are guaranteed to be notified in correct order at stream level under concurrent Equinox Transact calls.
     /// NOTE the caller should inspect and/or copy the event efficiently and immediately
     /// NOTE blocking and/or running reactions synchronously will hamper test performance and/or may result in deadlock
-    [<CLIEvent>] member _.Committed : IEvent<Commit<'Format>> = committed.Publish
+    [<CLIEvent>] member _.Committed: IEvent<Commit<'Format>> = committed.Publish
 
     /// Loads events from a given stream, null if none yet written
     member _.Load(streamName) =
@@ -38,7 +37,7 @@ type VolatileStore<'Format> private (lock) =
         events
 
     /// Attempts a synchronization operation - yields conflicting value if expectedCount does not match
-    member _.TrySync(streamName, _categoryName, _streamId, expectedCount, events) : struct (bool * FsCodec.ITimelineEvent<'Format>[]) =
+    member _.TrySync(streamName, _categoryName, _streamId, expectedCount, events): struct (bool * FsCodec.ITimelineEvent<'Format>[]) =
         // Where attempts overlap on the same stream, there's a race to raise the Committed event for each 'commit'
         // If we don't serialize the publishing of the events, its possible for handlers to observe the Events out of order
         // NOTE while a Channels based impl might offer better throughput at load, in practical terms serializing all Committed event notifications
@@ -61,32 +60,32 @@ type Token = int
 /// Internal implementation detail of MemoryStore
 module private Token =
 
-    let private streamTokenOfEventCount (eventCount : int) : StreamToken =
+    let private streamTokenOfEventCount (eventCount: int): StreamToken =
         // TOCONSIDER Could implement streamBytes tracking based on a supplied event size function (store is agnostic to format)
         { value = box eventCount; version = int64 eventCount; streamBytes = -1 }
-    let (|Unpack|) (token : StreamToken) : int = unbox<Token> token.value
+    let (|Unpack|) (token: StreamToken): int = unbox<Token> token.value
     /// Represent a stream known to be empty
     let ofEmpty = streamTokenOfEventCount 0
-    let ofValue (value : 'event[]) = streamTokenOfEventCount value.Length
+    let ofValue (value: 'event[]) = streamTokenOfEventCount value.Length
 
 /// Represents the state of a set of streams in a style consistent withe the concrete Store types - no constraints on memory consumption (but also no persistence!).
 type private Category<'event, 'state, 'context, 'Format>(store: VolatileStore<'Format>, codec: FsCodec.IEventCodec<'event, 'Format, 'context>, fold, initial) =
     interface ICategory<'event, 'state, 'context> with
-        member _.Load(_log, _categoryName, _streamId, streamName, _maxAge, _requireLeader, _ct) =
+        member _.Load(_log, _categoryName, _streamId, streamName, _maxAge, _requireLeader, _ct) = task {
             match store.Load(streamName) with
-            | null -> struct (Token.ofEmpty, initial) |> Task.FromResult
-            | xs -> struct (Token.ofValue xs, fold initial (Seq.chooseV codec.TryDecode xs)) |> Task.FromResult
-        member _.TrySync(_log, categoryName, streamId, streamName, context, _init, Token.Unpack eventCount, state, events, _ct) =
-            let inline map i (e : FsCodec.IEventData<'Format>) = FsCodec.Core.TimelineEvent.Create(int64 i, e)
+            | null -> return (Token.ofEmpty, initial)
+            | xs -> return (Token.ofValue xs, fold initial (Seq.chooseV codec.TryDecode xs)) }
+        member _.Sync(_log, categoryName, streamId, streamName, context, _init, Token.Unpack eventCount, state, events, _ct) = task {
+            let inline map i (e: FsCodec.IEventData<'Format>) = FsCodec.Core.TimelineEvent.Create(int64 i, e)
             let encoded = Array.ofSeq events |> Array.mapi (fun i e -> map (eventCount + i) (codec.Encode(context, e)))
             match store.TrySync(streamName, categoryName, streamId, eventCount, encoded) with
             | true, streamEvents' ->
-                SyncResult.Written (Token.ofValue streamEvents', fold state events) |> Task.FromResult
+                return SyncResult.Written (Token.ofValue streamEvents', fold state events)
             | false, conflictingEvents ->
-                let resync _ct =
+                let resync _ct = task {
                     let token' = Token.ofValue conflictingEvents
-                    struct (token', fold state (conflictingEvents |> Seq.skip eventCount |> Seq.chooseV codec.TryDecode)) |> Task.FromResult
-                SyncResult.Conflict resync |> Task.FromResult
+                    return struct (token', fold state (conflictingEvents |> Seq.skip eventCount |> Seq.chooseV codec.TryDecode)) }
+                return SyncResult.Conflict resync }
 
 type MemoryStoreCategory<'event, 'state, 'Format, 'context> internal (resolveInner, empty) =
     inherit Equinox.Category<'event, 'state, 'context>(resolveInner, empty)