+++ src/Equinox.SqlStreamStore/SqlStreamStore.fs
@@ -142,7 +142,7 @@ module private Write =
         try let! wr = conn.AppendToStream(StreamId streamName, (if version = -1L then ExpectedVersion.NoStream else int version), events, ct)
             return EsSyncResult.Written wr
         with :? WrongExpectedVersionException as ex ->
-            log.Information(ex, "SqlEs TrySync WrongExpectedVersionException writing {EventTypes}, expected {ExpectedVersion}",
+            log.Information(ex, "SqlEs Sync WrongExpectedVersionException writing {EventTypes}, expected {ExpectedVersion}",
                 [| for x in events -> x.Type |], version)
             return EsSyncResult.ConflictUnknown }
     let eventDataBytes events =
@@ -192,7 +192,7 @@ module private Read =
         : IAsyncEnumerable<int64 option * ResolvedEvent[]> =
         let rec loop batchCount pos: IAsyncEnumerable<int64 option * ResolvedEvent[]> = taskSeq {
             match maxPermittedBatchReads with
-            | Some mpbr when batchCount >= mpbr -> log.Information "batch Limit exceeded"; invalidOp "batch Limit exceeded"
+            | Some limit when batchCount >= limit -> log.Information "batch Limit exceeded"; invalidOp "batch Limit exceeded"
             | _ -> ()
 
             let batchLog = log |> Log.prop "batchIndex" batchCount
@@ -267,8 +267,8 @@ module private Read =
         let log = log |> Log.prop "batchSize" batchSize |> Log.prop "stream" streamName
         let startPosition = int64 Position.End
         let direction = Direction.Backward
-        let readlog = log |> Log.prop "direction" direction
-        let batchesBackward ct: IAsyncEnumerable<int64 option * ResolvedEvent[]> = readBatches readlog retryingLoggingReadSlice maxPermittedBatchReads startPosition ct
+        let readLog = log |> Log.prop "direction" direction
+        let batchesBackward ct: IAsyncEnumerable<int64 option * ResolvedEvent[]> = readBatches readLog retryingLoggingReadSlice maxPermittedBatchReads startPosition ct
         let! t, (version, events) = (batchesBackward >> mergeFromCompactionPointOrStartFromBackwardsStream log) |> Stopwatch.time ct
         log |> logBatchRead direction streamName t (Array.map ValueTuple.fst events) batchSize version
         return version, events }
@@ -308,12 +308,12 @@ module Token =
     let ofNonCompacting streamVersion: StreamToken =
         create None None streamVersion
     // headroom before compaction is necessary given the stated knowledge of the last (if known) `compactionEventNumberOption`
-    let private batchCapacityLimit compactedEventNumberOption unstoredEventsPending (batchSize: int) (streamVersion: int64): int =
+    let private batchCapacityLimit compactedEventNumberOption eventsPending (batchSize: int) (streamVersion: int64): int =
         match compactedEventNumberOption with
-        | Some (compactionEventNumber: int64) -> (batchSize - unstoredEventsPending) - int (streamVersion - compactionEventNumber + 1L) |> max 0
-        | None -> (batchSize - unstoredEventsPending) - (int streamVersion + 1) - 1 |> max 0
-    let (*private*) ofCompactionEventNumber compactedEventNumberOption unstoredEventsPending batchSize streamVersion: StreamToken =
-        let batchCapacityLimit = batchCapacityLimit compactedEventNumberOption unstoredEventsPending batchSize streamVersion
+        | Some (compactionEventNumber: int64) -> (batchSize - eventsPending) - int (streamVersion - compactionEventNumber + 1L) |> max 0
+        | None -> (batchSize - eventsPending) - (int streamVersion + 1) - 1 |> max 0
+    let (*private*) ofCompactionEventNumber compactedEventNumberOption eventsPending batchSize streamVersion: StreamToken =
+        let batchCapacityLimit = batchCapacityLimit compactedEventNumberOption eventsPending batchSize streamVersion
         create compactedEventNumberOption (Some batchCapacityLimit) streamVersion
     /// Assume we have not seen any compaction events; use the batchSize and version to infer headroom
     let ofUncompactedVersion batchSize streamVersion: StreamToken =
@@ -382,7 +382,7 @@ type SqlStreamStoreContext(connection: SqlStreamStoreConnection, batchOptions: B
             match events |> Array.tryFindBack (fun re -> match tryDecode re with ValueSome e -> isCompactionEvent e | _ -> false) with
             | None -> return Token.ofPreviousTokenAndEventsLength streamToken events.Length batchOptions.BatchSize version, Array.chooseV tryDecode events
             | Some resolvedEvent -> return Token.ofCompactionResolvedEventAndVersion resolvedEvent batchOptions.BatchSize version, Array.chooseV tryDecode events }
-    member internal _.TrySync(log, streamName, (Token.Unpack pos as streamToken), events, encodedEvents: EventData[], isCompactionEventType, ct): Task<GatewaySyncResult> = task {
+    member internal _.Sync(log, streamName, (Token.Unpack pos as streamToken), events, encodedEvents: EventData[], isCompactionEventType, ct): Task<GatewaySyncResult> = task {
         match! Write.writeEvents log connection.WriteRetryPolicy connection.WriteConnection streamName pos.streamVersion encodedEvents ct with
         | EsSyncResult.Written wr ->
             let version' = wr.CurrentVersion |> int64
@@ -435,7 +435,7 @@ type private Category<'event, 'state, 'context>(context: SqlStreamStoreContext, 
     interface ICategory<'event, 'state, 'context> with
         member _.Load(log, _categoryName, _streamId, streamName, _maxAge, requireLeader, ct) =
             fetch initial (loadAlgorithm log streamName requireLeader ct)
-        member _.TrySync(log, _categoryName, _streamId, streamName, ctx, _maybeInit, (Token.Unpack token as streamToken), state, events, ct) = task {
+        member _.Sync(log, _categoryName, _streamId, streamName, ctx, _maybeInit, (Token.Unpack token as streamToken), state, events, ct) = task {
             let events =
                 match access with
                 | None | Some AccessStrategy.LatestKnownEvent -> events
@@ -444,7 +444,7 @@ type private Category<'event, 'state, 'context>(context: SqlStreamStoreContext, 
                     if cc.IsCompactionDue then Array.append events (fold state events |> compact |> Array.singleton) else events
             let encode e = codec.Encode(ctx, e)
             let encodedEvents: EventData[] = events |> Array.map (encode >> UnionEncoderAdapters.eventDataOfEncodedEvent)
-            match! context.TrySync(log, streamName, streamToken, events, encodedEvents, compactionPredicate, ct) with
+            match! context.Sync(log, streamName, streamToken, events, encodedEvents, compactionPredicate, ct) with
             | GatewaySyncResult.Written token' ->  return SyncResult.Written  (token', fold state events)
             | GatewaySyncResult.ConflictUnknown -> return SyncResult.Conflict (reload (log, streamName, (*requireLeader*)true, streamToken, state)) }
     interface Caching.IReloadable<'state> with member _.Reload(log, sn, leader, token, state, ct) = reload (log, sn, leader, token, state) ct