+++ tests/Equinox.CosmosStore.Integration/DocumentStoreIntegration.fs
@@ -535,7 +535,7 @@ type Tests(testOutputHelper) =
         test <@ match res with
                 | Choice2Of2 e -> e.Message.StartsWith "Origin event not found; no Archive Container supplied"
                                   || e.Message.StartsWith "Origin event not found; no Archive Table supplied"
-                | x -> failwithf "Unexpected %A" x @>
+                | x -> failwithf $"Unexpected %A{x}" @>
         test <@ [EqxAct.ResponseForward; EqxAct.QueryForward] = capture.ExternalCalls @>
         verifyRequestChargesMax 3 // 2.99